Controle e Service Implementado

+++ Fature/src/main/java/br/com/capsule/model/Cbo.java
@@ -19,14 +19,14 @@ public class Cbo implements Serializable{
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String nome;
-	private int codigo;
+	private Long codigo;
 	
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getNome() {
@@ -35,10 +35,10 @@ public class Cbo implements Serializable{
 	public void setNome(String nome) {
 		this.nome = nome;
 	}
-	public int getCodigo() {
+	public Long getCodigo() {
 		return codigo;
 	}
-	public void setCodigo(int codigo) {
+	public void setCodigo(Long codigo) {
 		this.codigo = codigo;
 	}
